pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mbari-pbp"
version = "0.3.0b14"
# for now, pbp version "aligned" (not necessarily exactly) with that of lifewatch-pypam.
description = "PyPAM based Processing"
authors = [
    "Carlos Rueda <carueda@mbari.org>",
    "Danelle Cline <dcline@mbari.org>",
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mbari-org/pbp"
packages = [{include = "pbp"}]

[tool.poetry.scripts]
pbp = "pbp.main:main"
pbp-cloud = "pbp.main_cloud:main"
pbp-plot = "pbp.plot:main"
pbp-json-gen = "pbp.main_json_generator:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.12.0"
lifewatch-pypam = "0.3.0"
boto3 = "1.34.47"
google-cloud-storage = "2.14.0"
dataclasses-json = "0.6.4"
python-dateutil = "2.8.2"
pyyaml = "6.0.1"
progressbar2 = "3.53.1"
marshmallow = "3.20.2"
soundfile = "0.12.1"
h5netcdf = "1.3.0"
pvlib = "0.10.3"
Pyarrow = "15.0.0"
loguru = "0.7.2"

[tool.poetry.group.dev.dependencies]
mypy = "1.8.0"
ruff = "0.2.2"
pytest = "7.4.4"  # not upgraded due to conflict with pypam 0.3.0
syrupy = "4.6.1"
types-python-dateutil = "2.8.2"
types-PyYAML = "6.0.1"


[tool.mypy]
show_error_codes = true
ignore_missing_imports = true
exclude = [
    'examples/',
    'pbp/plotting\.py',
    'pbp/json_generator/metadata.*\.py',
    'pbp/json_generator/corrector\.py',
    'pbp/json_generator/gen_soundtrap\.py',
    'pbp/json_generator/gen_nrs\.py',
]

[tool.ruff]
line-length = 90
exclude = [
  "virtenv/**"
]

[tool.pytest.ini_options]
testpaths = ["tests",]
pythonpath = ["."]
